+++ CHANGELOG.md
@@ -2,15 +2,16 @@
 
 ### Added
  - Better MQTT input validation on Virtual Input, Output, Flag, Sensor and Counter 'set' values from MQTT.
-
  
 ### Breaking Changes ###
  
 ### Changed
  - Updated DOCS.md for ARM based system Battery Update/Refresh for future Slaves on ID 38 and 39.
  - Updated DOCS.md to better explain Zone count setting when using Virtual Zones.
+ - Increased Max. Zone Count from 96 to 128 for future expansion by Cytech.
    
 ### Fixed
+ - MQTT topic removal when Comfort Zones changed.
 
 
 ## [1.6.2]